fix(ci): 修复 clippy unnecessary_cast + python_sandbox flaky 超时
#184 (c277eb6) Lint 失败: cron_expr.rs 的 dt.minute() as u32 等冗余
cast 触发 clippy::unnecessary_cast (CI 用 -D warnings)。chrono 的
Timelike/Datelike minute()/hour()/day()/month() 已返回 u32,无需 cast。

#185 (1107a45) Test 失败: python_sandbox_blocks_socket_connect 因 Python
子进程超时(SKILL_TIMEOUT=5s) panic。Windows CI 上 Python 启动 +
SANDBOX_PREAMBLE 导入可能超过 5s。与同文件 python_sandbox_allows_local_io
一致, 将 Timeout 当作环境问题 skip 而非 panic。

验证: cargo clippy --features grpc,channels -- -D warnings 通过
      cargo fmt --all -- --check 通过

+++ src-tauri/src/evolution/cron_expr.rs
@@ -87,10 +87,12 @@ impl CronExpr {
     ///
     /// 这是 cron 的标准行为，确保 `"0 0 1 * 0"` 在每月 1 号**和**每周日都触发。
     pub fn matches(&self, dt: DateTime<Utc>) -> bool {
-        let min_match = self.minutes.contains(&(dt.minute() as u32));
-        let hour_match = self.hours.contains(&(dt.hour() as u32));
-        let dom_match = self.days_of_month.contains(&(dt.day() as u32));
-        let month_match = self.months.contains(&(dt.month() as u32));
+        // chrono 的 Timelike/Datelike minute()/hour()/day()/month() 已返回 u32,
+        // 无需 as u32 (clippy::unnecessary_cast)。
+        let min_match = self.minutes.contains(&dt.minute());
+        let hour_match = self.hours.contains(&dt.hour());
+        let dom_match = self.days_of_month.contains(&dt.day());
+        let month_match = self.months.contains(&dt.month());
         // chrono: Weekday::num_days_from_sunday() 返回 0=Sunday..=6=Saturday，与 cron 一致。
         let dow_match = self
             .days_of_week
